numbers: split the number operations in to different files
I can already see `numbers.go` getting huge, so let's try to keep the features
more manageable.
